move new function to right namespace
