+++ cpp/simulations/2020_npis_sarscov2_wildtype_germany.cpp
@@ -471,9 +471,9 @@ get_graph(mio::Date start_date, mio::Date end_date, const fs::path& data_dir)
     mio::Graph<mio::osecir::Model, mio::MigrationParameters> params_graph;
     const auto& read_function = mio::osecir::read_population_data_county<mio::osecir::Model>;
     const auto& create_graph_function =
-        mio::regions::create_graph<mio::osecir::TestAndTraceCapacity, mio::osecir::ContactPatterns, ContactLocation,
-                                   mio::osecir::InfectionState, mio::osecir::Model, mio::osecir::Parameters,
-                                   decltype(read_function)>;
+        mio::create_graph<mio::osecir::TestAndTraceCapacity, mio::osecir::ContactPatterns, ContactLocation,
+                          mio::osecir::InfectionState, mio::osecir::Model, mio::osecir::Parameters,
+                          decltype(read_function)>;
 
     BOOST_OUTCOME_TRY(create_graph_function(params_graph, params, start_date, end_date, data_dir, read_function,
                                             scaling_factor_infected, 1.0, 7.5 / 100000., migrating_compartments,
